Removes astropy dependency from code + tests
Simplifying things.

This commit removes a lot of boilerplate astropy stuff
that was not really used, and removes dependency on
astropy Table.

The core functionality of this ccsds reader does
not at all depend on astropy, astropy was simply
being depended on for the Table feature, which was
used to wrap the result from the interface.

Although this is a breaking change, the user should now
just wrap the return from ccsdspy.FixedLength.load with
astropy.table.Table in their own code if they are ok
with the astropy dependency.

Rationale: astropy is a mamoth of a dependency to
require in order to simply use astropy.table.Table.