Removes final mentions of astropy
Checking with `grep -r -i "astro" .` looks
as would be expected. The only current project
file with a mention of astro is now the CHANGELOG.

The new Sphinx Makefile and make.bat are taken
from a recent default produced by sphinx-quickstart
**the only modification to Makefile being to add
`rm -rf api` to the clean target, which was in the
old Makefile, in order to clean the api directory
created by sphinx-automodapi.